- Fixed some files to [include_what_you_use] principle - Renamed UpdateCoef to UpdateCoefficient - Ceilf to std::ceil

+++ Event/event.cpp
@@ -68,7 +68,7 @@ QDebug operator<<(QDebug debug, const Event& event) {
 
       case QMetaType::Float: {
         float value = arg.toFloat();
-        debug.nospace() << value << (ceilf(value) == value ? ".0f" : "f");
+        debug.nospace() << value << (std::ceil(value) == value ? ".0f" : "f");
         break;
       }
